Move UTF-8 string to own assembly. (#1603)
* Moving Utf8String into own assembly

* Removing an unused class from Utf8String assembly.

* CR Feedback
