+++ corefxlab.sln
@@ -140,6 +140,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.IO.Compression.Brotl
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.IO.Compression.Tests", "tests\System.IO.Compression.Tests\System.IO.Compression.Tests.csproj", "{74EEC686-FB85-43D5-95EA-10048C4A4567}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Text.Utf8String", "src\System.Text.Utf8String\System.Text.Utf8String.csproj", "{07CC9F28-FC67-453B-9057-AC9025E5CA67}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Text.Utf8String.Tests", "tests\System.Text.Utf8String.Tests\System.Text.Utf8String.Tests.csproj", "{C6F061D6-18D6-41E3-A692-C0603A994F30}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -906,6 +910,30 @@ Global
 		{74EEC686-FB85-43D5-95EA-10048C4A4567}.Release|x64.Build.0 = Release|Any CPU
 		{74EEC686-FB85-43D5-95EA-10048C4A4567}.Release|x86.ActiveCfg = Release|Any CPU
 		{74EEC686-FB85-43D5-95EA-10048C4A4567}.Release|x86.Build.0 = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|x64.Build.0 = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Debug|x86.Build.0 = Debug|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|x64.ActiveCfg = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|x64.Build.0 = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|x86.ActiveCfg = Release|Any CPU
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67}.Release|x86.Build.0 = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|x64.Build.0 = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Debug|x86.Build.0 = Debug|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|x64.ActiveCfg = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|x64.Build.0 = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|x86.ActiveCfg = Release|Any CPU
+		{C6F061D6-18D6-41E3-A692-C0603A994F30}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -974,5 +1002,7 @@ Global
 		{793E4C5F-33B9-44AA-BE44-54EE624E79CB} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{D57AB952-8596-4E4C-B310-3578AE538BA7} = {4B000021-5278-4F2A-B734-DE49F55D4024}
 		{74EEC686-FB85-43D5-95EA-10048C4A4567} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
+		{07CC9F28-FC67-453B-9057-AC9025E5CA67} = {4B000021-5278-4F2A-B734-DE49F55D4024}
+		{C6F061D6-18D6-41E3-A692-C0603A994F30} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 	EndGlobalSection
 EndGlobal
