+++ src/System.Buffers.Experimental/System.Buffers.Experimental.csproj
@@ -10,6 +10,7 @@
     <ProjectReference Include="..\System.Buffers.Primitives\System.Buffers.Primitives.csproj" />
     <ProjectReference Include="..\System.Collections.Sequences\System.Collections.Sequences.csproj" />
     <ProjectReference Include="..\System.Text.Primitives\System.Text.Primitives.csproj" />
+    <ProjectReference Include="..\System.Text.Utf8String\System.Text.Utf8String.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="System.Buffers" Version="$(CoreFxVersion)" />
