+++ src/System.Text.Primitives/System/Text/Encoding/Utf16/Utf16LittleEndianEncoder.cs
@@ -214,32 +214,5 @@ namespace System.Text.Utf16
         }
 
         #endregion Surrogate helpers
-
-        #region Old stuff
-
-        // TODO: Should we rewrite this to not use char.ConvertToUtf32 or is it fast enough?
-        public static bool TryDecodeCodePointFromString(string s, int index, out uint codePoint, out int encodedChars)
-        {
-            if (index < 0 || index >= s.Length)
-            {
-                codePoint = default(uint);
-                encodedChars = 0;
-                return false;
-            }
-
-            if (index == s.Length - 1 && char.IsSurrogate(s[index]))
-            {
-                codePoint = default(uint);
-                encodedChars = 0;
-                return false;
-            }
-
-            encodedChars = char.IsHighSurrogate(s[index]) ? 2 : 1;
-            codePoint = unchecked((uint)char.ConvertToUtf32(s, index));
-
-            return true;
-        }
-
-        #endregion Old stuff
     }
 }
