+++ src/System.Text.Primitives/System/Text/Encoding/Utf8/Utf8Encoder.cs
@@ -1031,74 +1031,5 @@ namespace System.Text.Utf8
         }
 
         #endregion Encoding implementation
-
-        #region Old stuff
-
-        #region Decoder
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool IsFirstCodeUnitInEncodedCodePoint(byte codeUnit)
-        {
-            return (codeUnit & Utf8NonFirstByteInCodePointMask) != Utf8NonFirstByteInCodePointValue;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool TryFindEncodedCodePointBytesCountGoingBackwards(ReadOnlySpan<byte> buffer, out int encodedBytes)
-        {
-            encodedBytes = 1;
-            ReadOnlySpan<byte> it = buffer;
-            // TODO: Should we have something like: Span<byte>.(Slice from the back)
-            for (; encodedBytes <= Utf8MaxCodeUnitsPerCodePoint; encodedBytes++, it = it.Slice(0, it.Length - 1))
-            {
-                if (it.Length == 0)
-                {
-                    encodedBytes = default(int);
-                    return false;
-                }
-
-                // TODO: Should we have Span<byte>.Last?
-                if (IsFirstCodeUnitInEncodedCodePoint(it[it.Length - 1]))
-                {
-                    // output: encodedBytes
-                    return true;
-                }
-            }
-
-            // Invalid unicode character or stream prematurely ended (which is still invalid character in that stream)
-            encodedBytes = default(int);
-            return false;
-        }
-
-        // TODO: Name TBD
-        // TODO: optimize?
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool TryDecodeCodePointBackwards(ReadOnlySpan<byte> buffer, out uint codePoint, out int encodedBytes)
-        {
-            if (TryFindEncodedCodePointBytesCountGoingBackwards(buffer, out encodedBytes))
-            {
-                int realEncodedBytes;
-                // TODO: Inline decoding, as the invalid surrogate check can be done faster
-                bool ret = TryDecodeCodePoint(buffer, buffer.Length - encodedBytes, out codePoint, out realEncodedBytes);
-                if (ret && encodedBytes != realEncodedBytes)
-                {
-                    // invalid surrogate character
-                    // we know the character length by iterating on surrogate characters from the end
-                    // but the first byte of the character has also encoded length
-                    // seems like the lengths don't match
-                    codePoint = default(uint);
-                    return false;
-                }
-
-                return true;
-            }
-
-            codePoint = default(uint);
-            encodedBytes = default(int);
-            return false;
-        }
-
-        #endregion
-
-        #endregion Old stuff
     }
 }
