+++ src/System.Text.Utf8String/Properties/AssemblyInfo.cs
@@ -1,0 +1,27 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTrademark("")]
+
+[assembly: InternalsVisibleTo("System.Text.Utf8String.Tests,PublicKey="+
+"0024000004800000940000000602000000240000525341310004000001000100039ac461fa5c82"+
+"c7dd2557400c4fd4e9dcdf7ac47e3d572548c04cd4673e004916610f4ea5cbf86f2b1ca1cb824f"+
+"2a7b3976afecfcf4eb72d9a899aa6786effa10c30399e6580ed848231fec48374e41b3acf88119"+
+"31343fc2f73acf72dae745adbcb7063cc4b50550618383202875223fc75401351cd89c44bf9b50"+
+"e7fa3796")]
+
+// Setting ComVisible to false makes the types in this assembly not visible
+// to COM components.  If you need to access a type in this assembly from
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("80e6332b-dc83-4dfc-9c19-8d89e988ba4e")]
