+++ src/System.Text.Utf8String/System/Text/Utf16LittleEndianCodePointEnumerable.cs
