+++ src/System.Text.Utf8String/System/Text/Utf8String.CodePointEnumerable.cs
