+++ src/System.Text.Utf8String/System/Text/Utf8String.CodePointEnumerator.cs
@@ -8,15 +8,14 @@ namespace System.Text.Utf8
 {
     partial struct Utf8String
     {
-        // TODO: Name TBD
-        public struct CodePointReverseEnumerator : IEnumerator<uint>, IEnumerator
+        public struct CodePointEnumerator : IEnumerator<uint>, IEnumerator
         {
             private ReadOnlySpan<byte> _buffer;
             private int _index;
             private int _currentLenCache;
-            private const int ResetIndex = -Utf8Encoder.Utf8MaxCodeUnitsPerCodePoint - 1;
+            private const int ResetIndex = -Utf8Helper.MaxCodeUnitsPerCodePoint - 1;
 
-            public unsafe CodePointReverseEnumerator(ReadOnlySpan<byte> buffer) : this()
+            public unsafe CodePointEnumerator(ReadOnlySpan<byte> buffer) : this()
             {
                 _buffer = buffer;
 
@@ -53,9 +52,8 @@ namespace System.Text.Utf8
                         throw new InvalidOperationException("Current does not exist");
                     }
 
-                    ReadOnlySpan<byte> buffer = _buffer.Slice(0, _index);
-                    uint ret;
-                    bool succeeded = Utf8Encoder.TryDecodeCodePointBackwards(buffer, out ret, out _currentLenCache);
+                    uint codePoint;
+                    bool succeeded = Utf8Helper.TryDecodeCodePoint(_buffer, _index, out codePoint, out _currentLenCache);
 
                     if (!succeeded || _currentLenCache == 0)
                     {
@@ -63,7 +61,7 @@ namespace System.Text.Utf8
                         throw new Exception("Invalid code point!");
                     }
 
-                    return ret;
+                    return codePoint;
                 }
             }
 
@@ -92,7 +90,7 @@ namespace System.Text.Utf8
                     }
                 }
 
-                _index -= _currentLenCache;
+                _index += _currentLenCache;
                 _currentLenCache = 0;
 
                 return HasValue();
@@ -101,7 +99,7 @@ namespace System.Text.Utf8
             // This is different than Reset, it goes to the first element not before first
             private void MoveToFirstPosition()
             {
-                _index = _buffer.Length;
+                _index = 0;
             }
 
             private bool IsOnResetPosition()
@@ -116,7 +114,7 @@ namespace System.Text.Utf8
                     return true;
                 }
 
-                return _index > 0;
+                return _index < _buffer.Length;
             }
 
             // This is different than MoveToFirstPosition, this actually goes before anything
