+++ src/System.Text.Utf8String/System/Text/Utf8String.Enumerator.cs
