+++ tests/System.Text.Primitives.Tests/Parsing/PrimitiveParserBoolTests.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Text.Utf8;
 using Xunit;
 
 namespace System.Text.Primitives.Tests
@@ -18,7 +17,7 @@ namespace System.Text.Primitives.Tests
         [InlineData("False1234", 7, false, 5)]
         public unsafe void BooleanPositiveTests(string text, int length, bool expectedValue, int expectedConsumed)
         {
-            byte[] byteBuffer = new Utf8String(text).CopyBytes();
+            byte[] byteBuffer = Text.Encoding.UTF8.GetBytes(text);
             ReadOnlySpan<byte> byteSpan = new ReadOnlySpan<byte>(byteBuffer);
             
             char[] charBuffer = text.ToCharArray();
